requirements.txt
tensorflow>=2.20.0
opencv-python>=4.8.1
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pillow>=10.0.0
albumentations>=1.3.0
streamlit>=1.28.0
flask>=2.3.0
pyyaml>=6.0.0
tqdm>=4.65.0
jupyter>=1.0.0